pjesa offers

+++ delete_offer.php
@@ -1,8 +1,18 @@
 <?php
+<?php
+session_start();
 include "db.php";
-$id = $_GET['id'];
 
+if (!isset($_SESSION['role']) || $_SESSION['role'] != 'admin') {
+    header("Location: offers.php");
+    exit();
+}
+
+$id = $_GET['id'];
 $stmt = $conn->prepare("DELETE FROM offers WHERE id=:id");
 $stmt->execute([":id" => $id]);
 
 header("Location: offers.php");
+exit();
+?>
+
